refactor(config): extract utilities to config_utils.py
Split config.py (1027 lines) into:
- config.py (799 lines) - RequirementsConfig class
- config_utils.py (255 lines) - helper functions

Extracted functions:
- matches_trigger() - tool invocation matching
- load_yaml_or_json() - config file I/O
- deep_merge() - dictionary merging
- write_local_config() - local config writing
- write_project_config() - project config writing

Also fixes duplicate deep_merge() in init_presets.py by importing
from config_utils instead.

Backwards compatible: config.py re-exports all utilities via __all__
so existing imports continue to work.
